Update main.py
Fixed a minor issue with the age verification. I documented the age has to be greater or equal to 1, but the age check is only satisfy by age being greater than 1. So now the age can be greater or equal to 1 and the age check will not throw an error.